# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quran Qaida Class Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UserProfile {
  id: string;
  email?: string | null;
  firstName?: string | null;
  lastName?: string | null;
  profileImageUrl?: string | null;
  role?: string | null;
}

export interface GetCurrentAuthUserResponse {
  user?: UserProfile;
  isAuthenticated: boolean;
}

export interface Student {
  id: number;
  userId: string;
  scheduleType: string;
  currentLesson: number;
  notes?: string | null;
  createdAt?: string;
  updatedAt?: string;
}

export type StudentWithUser = Student & {
  user?: UserProfile;
};

export interface CreateStudentBody {
  userId: string;
  scheduleType: string;
  currentLesson?: number;
  notes?: string;
}

export interface UpdateStudentBody {
  scheduleType?: string;
  currentLesson?: number;
  notes?: string;
}

export interface Topic {
  id: number;
  title: string;
  content?: string | null;
  createdAt?: string;
}

export interface Question {
  id: number;
  topicId: number;
  question: string;
  questionType: string;
  options?: string[] | null;
  correctAnswer: string;
  createdAt?: string;
}

export type TopicWithQuestions = Topic & {
  questions?: Question[];
};

export interface CreateTopicBody {
  title: string;
  content?: string;
}

export interface CreateQuestionBody {
  topicId: number;
  question: string;
  questionType: string;
  options?: string[];
  correctAnswer: string;
}

export interface Result {
  id: number;
  studentId: number;
  topicId: number;
  score: number;
  totalQuestions: number;
  date?: string;
}

export type ResultWithTopic = Result & {
  topic?: Topic;
};

export interface SubmitResultBody {
  studentId: number;
  topicId: number;
  score: number;
  totalQuestions: number;
}

export interface Recording {
  id: number;
  studentId: number;
  lesson: number;
  audioUrl: string;
  createdAt?: string;
}

export interface CreateRecordingBody {
  studentId: number;
  lesson: number;
  audioUrl: string;
}

export interface UploadResponse {
  url: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface ExchangeMobileAuthorizationCodeBody {
  code: string;
  code_verifier: string;
  redirect_uri: string;
  state: string;
  nonce?: string | null;
}

export interface ExchangeMobileAuthorizationCodeResponse {
  token: string;
}

export interface LogoutMobileSessionResponse {
  success: boolean;
}

export type UploadRecordingBody = {
  audio: Blob;
};