# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Quran Qaida Class Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /students/{id}/recordings:
    get:
      operationId: getStudentRecordings
      tags: [recordings]
      summary: Get recordings for a student
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of recordings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recording"
  /students/{id}/results:
    get:
      operationId: getStudentResults
      tags: [results]
      summary: Get test results for a student
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResultWithTopic"
  /recordings:
    post:
      operationId: createRecording
      tags: [recordings]
      summary: Create a new recording entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRecordingBody"
      responses:
        "201":
          description: Created recording
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recording"
  /recordings/upload:
    post:
      operationId: uploadRecording
      tags: [recordings]
      summary: Upload audio recording file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                audio:
                  type: string
                  format: binary
              required:
                - audio
      responses:
        "200":
          description: Upload result with URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResponse"
...
  /questions:
    post:
      operationId: createQuestion
      tags: [questions]
      summary: Create a question for a topic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuestionBody"
      responses:
        "201":
          description: Created question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
  /questions/{id}:
    put:
      operationId: updateQuestion
      tags: [questions]
      summary: Update a question
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuestionBody"
      responses:
        "200":
          description: Updated question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
    delete:
      operationId: deleteQuestion
      tags: [questions]
      summary: Delete a question
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /results:
    post:
      operationId: submitTestResult
      tags: [results]
      summary: Submit a test result
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitResultBody"
      responses:
        "201":
          description: Saved result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
  /mobile-auth/token-exchange:
    post:
      operationId: exchangeMobileAuthorizationCode
      tags: [auth]
      summary: Exchange mobile auth code for session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExchangeMobileAuthorizationCodeBody"
      responses:
        "200":
          description: Session token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExchangeMobileAuthorizationCodeResponse"
...
components:
  schemas:
...
    Topic:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
...
    Question:
      type: object
      properties:
        id:
          type: integer
        topicId:
          type: integer
        question:
          type: string
        questionType:
          type: string
        options:
          type: array
          items:
            type: string
          nullable: true
        correctAnswer:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - topicId
        - question
        - questionType
        - correctAnswer
    CreateQuestionBody:
      type: object
      properties:
        topicId:
          type: integer
        question:
          type: string
        questionType:
          type: string
        options:
          type: array
          items:
            type: string
        correctAnswer:
          type: string
      required:
        - topicId
        - question
        - questionType
        - correctAnswer
    Result:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        topicId:
          type: integer
        score:
          type: number
        totalQuestions:
          type: integer
        date:
          type: string
          format: date-time
      required:
        - id
        - studentId
        - topicId
        - score
        - totalQuestions
    ResultWithTopic:
      allOf:
        - $ref: "#/components/schemas/Result"
        - type: object
          properties:
            topic:
              $ref: "#/components/schemas/Topic"
    SubmitResultBody:
      type: object
      properties:
        studentId:
          type: integer
        topicId:
          type: integer
        score:
          type: number
        totalQuestions:
          type: integer
      required:
        - studentId
        - topicId
        - score
        - totalQuestions
    Recording:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        lesson:
          type: integer
        audioUrl:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - studentId
        - lesson
        - audioUrl
    CreateRecordingBody:
      type: object
      properties:
        studentId:
          type: integer
        lesson:
          type: integer
        audioUrl:
          type: string
      required:
        - studentId
        - lesson
        - audioUrl
    UploadResponse:
      type: object
      properties:
        url:
          type: string
      required:
        - url
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
    ExchangeMobileAuthorizationCodeBody:
      type: object
      properties:
        code:
          type: string
        code_verifier:
          type: string
        redirect_uri:
          type: string
        state:
          type: string
        nonce:
          type: string
          nullable: true
      required:
        - code
        - code_verifier
        - redirect_uri
        - state
    ExchangeMobileAuthorizationCodeResponse:
      type: object
      properties:
        token:
          type: string
      required:
        - token
...