# lib/api-zod/src/generated/types/createQuestionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quran Qaida Class Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateQuestionBody {
  topicId: number;
  question: string;
  questionType: string;
  options?: string[];
  correctAnswer: string;
}