# lib/api-zod/src/generated/types/resultWithTopic.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quran Qaida Class Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { Result } from "./result";
import type { Topic } from "./topic";

export type ResultWithTopic = Result & {
  topic?: Topic;
};